backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
